Introduce prompt variations in reaction prompt
A problem the first week was that responses were too similar.
I don't really want to give the bot access to the history of old
reactions, as that introduces complexity. To add more variation,
I'm trying to include a randomly chosen reaction variation every
day. More may be added if necessary.